vbos-backend/requirements.txt
# Core
Django==5.2.5
django-configurations==2.5.1
gunicorn==23.0.0
setuptools==78.1.1

# For the persistence stores
psycopg2-binary==2.9.10
dj-database-url==2.3.0

# Model Tools
django-model-utils==5.0.0
django_unique_upload==0.2.1

# Rest apis
djangorestframework==3.16.1
djangorestframework-gis==1.2.0
django-filter==25.1
drf_spectacular==0.28.0
django-cors-headers==4.7.0
drf-excel==2.5.3

# Storage
django-storages==1.14.6
boto3==1.40.26
whitenoise

# Developer Tools
ipdb==0.13.13
ipython==8.30.0
mkdocs==1.6.1
flake8==7.1.1

# Testing
mock==5.1.0
factory-boy==3.3.3
pytest-django==4.9.0
coverage==7.6.9
